+++ crates/paksmith-core/src/error.rs
@@ -887,8 +887,12 @@ pub enum EncodedFault {
         /// strictly less than `file_count` (equality is the valid
         /// case and short-circuits the error; the per-push guard
         /// ensures the `>` case is caught earlier as
-        /// [`Self::FdiFileCountExceeded`]).
-        actual: u32,
+        /// [`Self::FdiFileCountExceeded`]). `u64` (not `u32`) so a
+        /// future invariant-breaking refactor that lets
+        /// `entries.len()` exceed `u32::MAX` can't silently truncate
+        /// to a value matching `file_count` and produce a self-
+        /// contradictory diagnostic. Issue #136.
+        actual: u64,
     },
 }
